requirements.txt
dj-database-url==0.4.1
Django==1.11.1
djangorestframework
google-api-python-client
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
requests
apiai
python-dateutil
Pyrebase
oauth2client
